Cast to signed int to squelch compiler warning

+++ Modem/compression/heatshrink_encoder.c
@@ -527,7 +527,7 @@ static uint16_t find_longest_match(heatshrink_encoder *hse, uint16_t start,
         pos = hsi->index[pos];
     }
 #else    
-    for (int16_t pos=end - 1; pos >= start; pos--) {
+    for (int16_t pos=end - 1; pos >= (int16_t)start; pos--) {
         for (len=0; len<maxlen; len++) {
             if (0) {
                 LOG("  --> cmp buf[%d] == 0x%02x against %02x (start %u)\n",
